Remove keep_lines from Source
Source was using keep_lines to remove the last read line when reading the next,
unless keep_lines was true. This was used to keep all source in memory if
the user requests a list file, or delete them otherwise.

The removal of the last line moved to the scanner after the line is fully
parsed. This solves the issue that when recursing, e.g. to parse an INCLUDE,
the original line was deleted and was not present when the INCLUDE returned.